Refactor user and key commands
