+++ src/bot/events/handlers/handleContextMenu.ts
@@ -1,0 +1,21 @@
+import { ContextMenuCommandInteraction } from "discord.js";
+import { SlashCommand } from "../../../../types";
+import handleCommandError from "./handleCommandError";
+
+export default async function handleContextMenu(
+  interaction: ContextMenuCommandInteraction
+) {
+  const context: SlashCommand | undefined =
+    interaction.client.commands.get(interaction.commandName);
+  if (!context || !context.enable) {
+    return interaction.reply(
+      "This feature is not available at the moment, please try again later."
+    );
+  }
+
+  try {
+    await context.execute(interaction);
+  } catch (error: any) {
+    await handleCommandError(interaction, error);
+  }
+}
